backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.0
loguru==0.7.2
python-dotenv==1.0.0
websockets==12.0
google-generativeai==0.3.2
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1
python-dotenv==1.0.0
